Unset check for foreign keys on down()

+++ database/migrations/2019_01_21_002915_create_project_tag_table.php
@@ -13,14 +13,16 @@ class CreateProjectTagTable extends Migration
      */
     public function up()
     {
-        Schema::create('project_tag', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('project_id');
-            $table->unsignedInteger('tag_id');
-            $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if(!Schema::hasTable('project_tag')) {
+            Schema::create('project_tag', function (Blueprint $table) {
+                $table->increments('id');
+                $table->unsignedInteger('project_id');
+                $table->unsignedInteger('tag_id');
+                $table->foreign('project_id')->references('id')->on('projects')->onDelete('cascade');
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -30,6 +32,8 @@ class CreateProjectTagTable extends Migration
      */
     public function down()
     {
+        DB::statement('SET FOREIGN_KEY_CHECKS = 0');
         Schema::dropIfExists('projects_tag');
+        DB::statement('SET FOREIGN_KEY_CHECKS = 1');
     }
 }
